ticket_system/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
twilio==9.8.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1